pyproject.toml
[project]
name = "montoring-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.0.0",
    "beanie>=2.0.0",
    "beautifulsoup4>=4.14.2",
    "celery[redis]>=5.5.3",
    "fake-useragent>=2.2.0",
    "fastapi[standard]>=0.119.1",
    "google-generativeai>=0.8.3",
    "httpx>=0.28.1",
    "langgraph>=1.0.1",
    "linkedin-scraper>=2.11.5",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymongo[srv]>=4.15.3",
    "redis>=5.2.1",
    "selenium>=4.37.0",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]
